+++ ximpol/irf/base.py
@@ -17,6 +17,10 @@
 # 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
 
+"""Basic format definitions related to the instrument response functions.
+"""
+
+
 import time
 
 from  astropy.io import fits
@@ -24,13 +28,36 @@ from  astropy.io import fits
 from ximpol.__version__ import TAG
 
 
+"""Specifications for the FITS headers related to the OGIP standards.
+(These are in common to all the fits files.)
+"""
+
+OGIP_HEADER_SPECS = [
+    ('HDUCLASS', 'OGIP    ', 'format conforms to OGIP standard'),
+    ('HDUVERS' , '1.1.0   ', 'Version of format (OGIP memo CAL/GEN/92-002a)'),
+    ('HDUDOC'  , 'OGIP memos CAL/GEN/92-002 & 92-002a', 'Documents describing the forma'),
+    ('HDUVERS1', '1.0.0   ', 'Obsolete - included for backwards compatibility'),
+    ('HDUVERS2', '1.1.0   ', 'Obsolete - included for backwards compatibility')
+]
+
+
+
 class xPrimaryHDU(fits.PrimaryHDU):
 
-    """
+    """Class describing a primary HDU to be written in a FITS file.
+
+    This is initializing a standard astropy.io.fits.PrimaryHDU object and
+    adding the creator and creation time fields. (The constructor takes no
+    arguments.)
+
+    Examples
+    --------
+    >>> primary_hdu = xPrimaryHDU()
+    >>> print(repr(primary_hdu.header))
     """
 
     def __init__(self):
-        """
+        """Constructor.
         """
         fits.PrimaryHDU.__init__(self)
         creator = 'ximpol %s' % TAG
@@ -44,16 +71,38 @@ class xPrimaryHDU(fits.PrimaryHDU):
 
 class xColDefsBase(fits.ColDefs):
 
-    """
+    """Column definitions base class.
+
+    This is a do-nothing class. Subclasses must define the COLUMN_SPECS
+    class member in the form of a list of three-elements tuples containing
+    (in this order):
+
+    * the name of the column;
+    * the format of the column content;
+    * the physical units of the column, e.g.:
+
+    >>> COLUMN_SPECS = [
+    >>>       ('ENERG_LO', 'E', 'keV'),
+    >>>       ('ENERG_HI', 'E', 'keV'),
+    >>>       ('SPECRESP', 'E', 'cm**2')
+    >>> ]
+
+    The actual data are filled in at creation time.
+
+    Args
+    ----
+    data : list of arrays
+        The actual data to be put in the columns. (The number and shapes of the
+        arrays have to match the column definitions.)
     """
 
     COLUMN_SPECS = []
 
-    def __init__(self, columns):
+    def __init__(self, data):
         """
         """
         cols = []
-        for i, (name, fmt) in enumerate(self.COLUMN_SPECS):
-            col = fits.Column(name = name, format = fmt, array = columns[i])
+        for i, (name, fmt, units) in enumerate(self.COLUMN_SPECS):
+            col = fits.Column(name, fmt, units, array = data[i])
             cols.append(col)
         fits.ColDefs.__init__(self, cols)