+++ ximpol/irf/mrf.py
@@ -17,35 +17,27 @@
 # 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
 
-from ximpol.irf.base import xColDefsBase
+from ximpol.irf.base import xColDefsBase, OGIP_HEADER_SPECS
 
+
+"""Header specifications for the MODFRESP extension of .mrf FITS files.
+"""
 
 MODFRESP_HEADER_SPECS = [
-    ('TUNIT1'  , 'keV     ', 'physical unit of field'),
-    ('TUNIT2'  , 'keV     ', 'physical unit of field'),
-    ('TUNIT3'  , '        ', 'physical unit of field'),
     ('EXTNAME' , 'MODFRESP', 'name of this binary table extension'),
-    ('HDUCLASS', 'OGIP    ', 'format conforms to OGIP standard'),
     ('HDUCLAS1', 'RESPONSE', 'dataset relates to spectral response'),
-    ('HDUCLAS2', 'MODFRESP', 'dataset contains spectral response'),
-    ('HDUVERS' , '1.1.0   ', 'Version of format (OGIP memo CAL/GEN/92-002a)'),
-    ('HDUDOC'  , 'OGIP memos CAL/GEN/92-002 & 92-002a', 'Documents describing the forma'),
-    ('HDUVERS1', '1.0.0   ', 'Obsolete - included for backwards compatibility'),
-    ('HDUVERS2', '1.1.0   ', 'Obsolete - included for backwards compatibility'),
-    ('FILTER'  , 'NONE'      , 'filter in use'),
-    ('PHAFILE' , 'UNKNOWN ', 'PHA file for which this ARF created'),
-    ('CREATOR' , 'NONE'    , 's/w task which wrote this dataset'),
-    ('RESPFILE', 'NONE'    , '')
-]
+    ('HDUCLAS2', 'MODFRESP', 'dataset contains modulation response')
+] + OGIP_HEADER_SPECS
 
 
 class xColDefsMODFRESP(xColDefsBase):
 
-    """
+    """ximpol.irf.base.xColDefsBase subclass for the MODFRESP extension
+    of .mrf FITS files.
     """
 
     COLUMN_SPECS = [
-        ('ENERG_LO', 'E'),
-        ('ENERG_HI', 'E'),
-        ('MODFRESP', 'E')
+        ('ENERG_LO', 'E', 'keV'),
+        ('ENERG_HI', 'E', 'keV'),
+        ('MODFRESP', 'E', None)
     ]